```
from asv_bench.benchmarks.io.style import *
self = Render()
self.setup(36, 12)

%timeit self.setup(36, 12); self.time_apply_render(36, 12)
29.2 ms ± 1.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <- master
9.98 ms ± 706 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <- PR
```

cc @attack68