@attack68 2021-08-29
nice. good spot. lgtm. tests seem unrelated. there might be one or two other places this pattern might be better too.